Reduces controller to more manageable size and extracts functionality to helpers
